Remove Pydantic from core dependencies
